+++ packages/brick_offline_first_with_rest_build/test/offline_first_schema_generator/test_with_associations.dart
@@ -31,7 +31,7 @@ final schema = Schema(0,
             SchemaColumn('_brick_id', int,
                 autoincrement: true, nullable: false, isPrimaryKey: true),
             SchemaColumn('assoc_SqliteAssoc_brick_id', int,
-                isForeignKey: true, foreignTableName: 'SqliteAssoc'),
+                isForeignKey: true, foreignTableName: 'SqliteAssoc', onCascadeDelete: false),
             SchemaColumn('assocs', String)
           ]))
     ]));