+++ packages/brick_sqlite_abstract/lib/src/db/migration_commands/insert_foreign_key.dart
@@ -10,10 +10,15 @@ class InsertForeignKey extends MigrationCommand {
   /// Defaults to lowercase `${foreignTableName}_brick_id`
   final String foreignKeyColumn;
 
+  /// When true, deletion of a row within this model's table will delete all
+  /// referencing children records. Defaults `false`.
+  final bool onDeleteCascade;
+
   const InsertForeignKey(
     this.localTableName,
     this.foreignTableName, {
     this.foreignKeyColumn,
+    this.onDeleteCascade = false,
   });
 
   String get _foreignKeyColumn {
@@ -24,13 +29,15 @@ class InsertForeignKey extends MigrationCommand {
     return foreignKeyColumnName(foreignTableName);
   }
 
+  String get _cascadeStatement => onDeleteCascade ? ' ON DELETE CASCADE' : '';
+
   @override
   String get statement =>
-      'ALTER TABLE `$localTableName` ADD COLUMN `$_foreignKeyColumn` INTEGER REFERENCES `$foreignTableName`(`${InsertTable.PRIMARY_KEY_COLUMN}`)';
+      'ALTER TABLE `$localTableName` ADD COLUMN `$_foreignKeyColumn` INTEGER REFERENCES `$foreignTableName`(`${InsertTable.PRIMARY_KEY_COLUMN}`)$_cascadeStatement';
 
   @override
   String get forGenerator =>
-      "InsertForeignKey('$localTableName', '$foreignTableName', foreignKeyColumn: '$_foreignKeyColumn')";
+      "InsertForeignKey('$localTableName', '$foreignTableName', foreignKeyColumn: '$_foreignKeyColumn', onDeleteCascade: $onDeleteCascade)";
 
   @override
   MigrationCommand get down => DropColumn(_foreignKeyColumn, onTable: localTableName);