+++ packages/brick_sqlite_generators/lib/src/sqlite_schema/sqlite_schema_generator.dart
@@ -128,6 +128,7 @@ class SqliteSchemaGenerator {
         isForeignKey: true,
         foreignTableName: checker.unFuturedType.getDisplayString(),
         nullable: column?.nullable,
+        onDeleteCascade: onDeleteCascade,
       );
     } else if (checker.isMap || checker.isIterable) {
       // Iterables and Maps are stored as JSON